feat: reorganize repo
