+++ test/cosmoz-tabs-basic.test.js
@@ -3,7 +3,7 @@ import {
 } from '@open-wc/testing';
 import { spy } from 'sinon';
 
-import '../cosmoz-tabs.js';
+import '../src/cosmoz-tabs.js';
 
 suite('cosmoz-tabs', () => {
 	let tabs;
