+++ web-dev-server.config.mjs
@@ -1,0 +1,6 @@
+import { storybookPlugin } from '@web/dev-server-storybook';
+
+export default {
+	plugins: [storybookPlugin({ type: 'web-components' })],
+	nodeResolve: true
+};
